Indicate in admin if tasks building. Closes #85
A message indicating tasks are still building is now displayed to
challenge owners viewing a challenge that is in a building status. A
button is provided for refreshing the status, along with an indication
of how long it's been since the status was last refreshed.
