+++ src/components/AdminPane/Manage/ViewChallenge/ViewChallenge.js
@@ -1,39 +1,24 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import classNames from 'classnames'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import _get from 'lodash/get'
-import _map from 'lodash/map'
 import { Link } from 'react-router-dom'
-import { TaskStatus,
-         keysByStatus,
-         messagesByStatus } from '../../../../services/Task/TaskStatus/TaskStatus'
-import { MAPBOX_LIGHT,
-         layerSourceWithName }
-       from '../../../../services/VisibleLayer/LayerSources'
 import WithCurrentChallenge
        from '../../HOCs/WithCurrentChallenge/WithCurrentChallenge'
 import WithFilteredClusteredTasks
        from '../../HOCs/WithFilteredClusteredTasks/WithFilteredClusteredTasks'
-import WithBoundedTasks
-       from '../../HOCs/WithBoundedTasks/WithBoundedTasks'
 import Sidebar from '../../../Sidebar/Sidebar'
 import CommentList from '../../../CommentList/CommentList'
-import MapPane from '../../../EnhancedMap/MapPane/MapPane'
-import ChallengeTaskMap from '../ChallengeTaskMap/ChallengeTaskMap'
-import TaskAnalysisTable from '../TaskAnalysisTable/TaskAnalysisTable'
 import ChallengeOverview from '../ManageChallenges/ChallengeOverview'
 import BusySpinner from '../../../BusySpinner/BusySpinner'
 import SvgSymbol from '../../../SvgSymbol/SvgSymbol'
 import Tabs from '../../../Bulma/Tabs'
 import ChallengeMetrics from '../ChallengeMetrics/ChallengeMetrics'
 import ConfirmAction from '../../../ConfirmAction/ConfirmAction'
+import ViewChallengeTasks from './ViewChallengeTasks'
 import manageMessages from '../Messages'
 import messages from './Messages'
 import './ViewChallenge.css'
-
-const BoundedTaskTable =
-  WithBoundedTasks(TaskAnalysisTable, 'filteredClusteredTasks', 'taskInfo')
 
 /**
  * ViewChallenge displays various challenge details and metrics of interest
@@ -42,16 +27,6 @@ const BoundedTaskTable =
  * @author [Neil Rotstan](https://github.com/nrotstan)
  */
 export class ViewChallenge extends Component {
-  state = {
-    mapBounds: null,
-    mapZoom: null,
-    renderingProgress: null,
-  }
-
-  /** Invoked by the map when the user pans or zooms */
-  updateMapBounds = (challengeId, bounds, zoom) =>
-    this.setState({mapBounds: bounds, mapZoom: zoom})
-
   deleteChallenge = () => {
     this.props.deleteChallenge(this.props.challenge.parent.id,
                                this.props.challenge.id)
@@ -72,36 +47,6 @@ export class ViewChallenge extends Component {
 
       [this.props.intl.formatMessage(messages.challengeMetricsTabLabel)]:
         <ChallengeMetrics challenges={[this.props.challenge]} />,
-    }
-
-    // Use CSS Modules once supported by create-react-app
-    const statusColors = {
-      [TaskStatus.created]: '#0082C8',       // $status-created-color
-      [TaskStatus.fixed]: '#3CB44B',         // $status-fixed-color
-      [TaskStatus.falsePositive]: '#F58231', // $status-falsePositive-color
-      [TaskStatus.skipped]: '#FFE119',       // $status-skipped-color
-      [TaskStatus.deleted]: '#46F0F0',       // $status-deleted-color
-      [TaskStatus.alreadyFixed]: '#911EB4',  // $status-alreadyFixed-color
-      [TaskStatus.tooHard]: '#E6194B',       // $status-tooHard-color
-    }
-
-    const statusFilters = _map(TaskStatus, status => (
-      <div key={status} className="status-filter is-narrow">
-        <div className={classNames("field", keysByStatus[status])}
-             onClick={() => this.props.toggleIncludedStatus(status)}>
-          <input className="is-checkradio is-circle has-background-color is-success" type="checkbox"
-                 checked={this.props.includeStatuses[status]}
-                 onChange={() => null} />
-          <label>
-            <FormattedMessage {...messagesByStatus[status]} />
-          </label>
-        </div>
-      </div>
-    ))
-
-    const filterOptions = {
-      includeStatuses: this.props.includeStatuses,
-      withinBounds: this.state.mapBounds,
     }
 
     return (
@@ -153,26 +98,7 @@ export class ViewChallenge extends Component {
           </Sidebar>
 
           <div className="admin__manage__primary-content">
-            <div className='admin__manage-tasks'>
-              <div className="status-filter-options">
-                {statusFilters}
-              </div>
-
-              <MapPane>
-                <ChallengeTaskMap taskInfo={this.props.filteredClusteredTasks}
-                                  setChallengeMapBounds={this.updateMapBounds}
-                                  lastBounds={this.state.mapBounds}
-                                  lastZoom={this.state.mapZoom}
-                                  statusColors={statusColors}
-                                  filterOptions={filterOptions}
-                                  monochromaticClusters
-                                  defaultLayer={layerSourceWithName(MAPBOX_LIGHT)}
-                                  {...this.props} />
-              </MapPane>
-              <BoundedTaskTable filterOptions={filterOptions}
-                                totalTaskCount={_get(this.props, 'clusteredTasks.tasks.length')}
-                                {...this.props} />
-            </div>
+            <ViewChallengeTasks {...this.props} />
           </div>
         </div>
       </div>
