+++ src/components/EnhancedMap/MapPane/MapPane.scss
@@ -49,7 +49,7 @@
       border-radius: $radius-medium;
 
       .leaflet-popup-content {
-        width: 300px;
+        width: 350px;
         overflow-y: auto;
         margin: 0;
 
@@ -57,6 +57,8 @@
           color: $grey;
           font-size: $size-7;
           margin: 19px 0px;
+          max-height: 70vh;
+          overflow-y: auto;
 
           h3 {
             color: $primary;
